call-channel: Remove the CallState interface
The CallState interface is not meant for Call changes, all the
information needed is in the remote call member flags and the call
flags. As a first step remote CallState as it breaks some signals from
the Call interface and add TODO items for reimplementing things
correctly using the new spec work.